requirements.txt
fastapi
uvicorn
torch
transformers
accelerate
python-dotenv
compressed-tensors